Data annotation for each domain

+++ Customer.Inquiry.Domain/Interface/ITransaction.cs
@@ -4,7 +4,7 @@ namespace Customer.Inquiry.Domain.Interface
 {
     public interface ITransaction : IDomainBase
     {
-        float Amount { get; set; }
+        decimal Amount { get; set; }
         string CurrencyCode { get; set; }
         TransactionStatus Status { get; set; }
     }
